UPGRADE: CLICK TO CLIPBOARD, OVERFLOW, DESCRIPTION
Those are the optional features added:
- Clicking on a node or key/value while copy its content:
    > Clicking on a node will copy the object to the user's clipboard
    > Clicking on a key/value will copy the value
 - Auto overflow: when the mouse overs a value, it sets: overflow=auto
    > Useful if the value goes outside of the screen
 - Set a description above the json tree

Had to rename CSS classes due to a conflict with other frameworks.

Notes:
The click-to-clipboard feature works well but has a minor limitation:
If a key/node's name appears twice in the same json: possible mistake.
This can be corrected fairly easily.
But I do not have the time currently.